Fix content? method error
Previously, @s3resource.object(@folder) cannot determine whether or not the @folder / directory exist. If called exists? on a path object, it will not exist despite the path existing with content in them. Fix added which now allow the resource to quickly summarize and determine if the directory has any files at all. If it doesn't have any files, then the path directory is automatically deleted