Change spec_helper test environment configuration variables
Before, the s3 test were relying on manual manipulation of the env configurations to run it tests. It was quickly shown to be bad code, because it is basically brute-forcing a solution in which still made changes to the original config and metadata. This change will allow configuration to be fetch from a custom config file in tmp/test_base.

Add TEST_CONFIG_PATH which points to the temp config file created.