Make UpstreamSource intermediate specs from upstream
As discovered in #336, the retirement of the v1 dependencies API revealed that UpstreamSource, when the dependencies API is unavailable, tells the client to go talk to rubygems directly for the spec files. This model does not work in general, because there is no guarantee that the client can access rubygems directly. This commit changes UpstreamSource to intermediate the connection