authorize: support naming and listing authorizations
Over time, gemstash authorizations turn into a pile of opaque tokens and
associated permissions. Keeping track of which token belongs to which
developer or automated system is required to enabled proper revocation
or rotation. Today, that requires the use of an external system, which
has the added problem of duplicating the key itself.

Support keeping this association within Gemstash itself by allowing for
authorizations to be named. Also support listing authorizations, so it's
easy to determine which key needs to be revoked, rotated, or otherwise
updated. See the manpage updates for details on use.

Resolve #364

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>